backend/requirements.txt
flask
flask_login
flask_cors
flask_sqlalchemy
python-dotenv==0.19.0
openai>=1.73.0